# openapi/v1.yaml
...
components:
  schemas:
...
    KeyDetailUsage:
      description: Aggregate usage embedded in the single-key detail response.
      properties:
        daily_tokens:
          title: Daily Tokens
          type: integer
        monthly_cost_usd:
          title: Monthly Cost Usd
          type: number
        monthly_tokens:
          title: Monthly Tokens
          type: integer
      required:
      - daily_tokens
      - monthly_tokens
      - monthly_cost_usd
      title: KeyDetailUsage
      type: object
...
    KeyUsageDailyEntry:
      description: A single day's usage row within the key-usage response.
      properties:
        cost_usd:
          anyOf:
          - type: number
          - type: 'null'
          title: Cost Usd
        date:
          title: Date
          type: string
        request_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Request Count
        tokens_input:
          anyOf:
          - type: integer
          - type: 'null'
          title: Tokens Input
        tokens_output:
          anyOf:
          - type: integer
          - type: 'null'
          title: Tokens Output
        waddleai_tokens:
          anyOf:
          - type: integer
          - type: 'null'
          title: Waddleai Tokens
      required:
      - date
      - waddleai_tokens
      - tokens_input
      - tokens_output
      - request_count
      - cost_usd
      title: KeyUsageDailyEntry
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  title: WaddleAI Management API
  version: '1'
openapi: 3.1.0
paths:
...
  /api/v1/keys:
    get:
      description: ''
      operationId: get_list_keys
      parameters: []
      responses:
        '200':
          content:
            application/json:
              schema:
                description: Response body for GET /api/v1/keys.
                properties:
                  keys:
                    items:
                      $ref: '#/components/schemas/KeySummary'
                    title: Keys
                    type: array
                  total:
                    title: Total
                    type: integer
                required:
                - keys
                - total
                title: KeyListResponse
                type: object
          description: Response body for GET /api/v1/keys.
      security:
      - bearerAuth: []
      summary: List virtual keys based on user role
      tags:
      - Keys
    post:
      description: ''
      operationId: post_create_key
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              description: Request body for POST /api/v1/keys.
              properties:
                allowed_models:
                  anyOf:
                  - items:
                      type: string
                    type: array
                  - type: 'null'
                  default: null
                  title: Allowed Models
                allowed_providers:
                  anyOf:
                  - items:
                      type: string
                    type: array
                  - type: 'null'
                  default: null
                  title: Allowed Providers
                budget_limit_daily:
                  anyOf:
                  - type: number
                  - type: 'null'
                  default: null
                  title: Budget Limit Daily
                budget_limit_monthly:
                  anyOf:
                  - type: number
                  - type: 'null'
                  default: null
                  title: Budget Limit Monthly
                expires_days:
                  anyOf:
                  - type: integer
                  - type: 'null'
                  default: 365
                  title: Expires Days
                name:
                  anyOf:
                  - type: string
                  - type: 'null'
                  default: null
                  title: Name
                organization_id:
                  anyOf:
                  - type: integer
                  - type: 'null'
                  default: null
                  title: Organization Id
                rpm_limit:
                  anyOf:
                  - type: integer
                  - type: 'null'
                  default: 60
                  title: Rpm Limit
                tpm_limit:
                  anyOf:
                  - type: integer
                  - type: 'null'
                  default: 10000
                  title: Tpm Limit
                user_id:
                  anyOf:
                  - type: integer
                  - type: 'null'
                  default: null
                  title: User Id
              title: CreateKeyRequest
              type: object
      responses:
        '201':
          content:
            application/json:
              schema:
                description: 'Response body for a successful POST /api/v1/keys.


                  The plaintext `api_key` is only ever returned here and on rotate
                  --

                  it is not recoverable afterwards, only `key_prefix` is persisted
                  for display.'
                properties:
                  api_key:
                    title: Api Key
                    type: string
                  expires_at:
                    anyOf:
                    - type: string
                    - type: 'null'
                    title: Expires At
                  id:
                    title: Id
                    type: integer
                  key_prefix:
                    title: Key Prefix
                    type: string
                  message:
                    title: Message
                    type: string
                  name:
                    title: Name
                    type: string
                required:
                - id
                - name
                - api_key
                - key_prefix
                - expires_at
                - message
                title: CreateKeyResponse
                type: object
          description: 'Response body for a successful POST /api/v1/keys.


            The plaintext `api_key` is only ever returned here and on rotate --

            it is not recoverable afterwards, only `key_prefix` is persisted for display.'
      security:
      - bearerAuth: []
      summary: Create a new virtual key
      tags:
      - Keys
  /api/v1/keys/{key_id}:
    delete:
      description: ''
      operationId: delete_delete_key
      parameters:
      - in: path
        name: key_id
        required: true
        schema:
          type: number
      responses:
        '200':
          content:
            application/json:
              schema:
                description: 'Generic `{"message": str}` envelope used by several
                  key endpoints.'
                properties:
                  message:
                    title: Message
                    type: string
                required:
                - message
                title: MessageResponse
                type: object
          description: 'Generic `{"message": str}` envelope used by several key endpoints.'
      security:
      - bearerAuth: []
      summary: Revoke/delete virtual key
      tags:
      - Keys
    get:
      description: ''
      operationId: get_get_key
      parameters:
      - in: path
        name: key_id
        required: true
        schema:
          type: number
      responses:
        '200':
          content:
            application/json:
              schema:
                description: Response body for GET /api/v1/keys/<key_id>.
                properties:
                  allowed_models:
                    anyOf:
                    - items:
                        type: string
                      type: array
                    - type: 'null'
                    title: Allowed Models
                  allowed_providers:
                    anyOf:
                    - items:
                        type: string
                      type: array
                    - type: 'null'
                    title: Allowed Providers
                  budget_limit_daily:
                    anyOf:
                    - type: number
                    - type: 'null'
                    title: Budget Limit Daily
                  budget_limit_monthly:
                    anyOf:
                    - type: number
                    - type: 'null'
                    title: Budget Limit Monthly
                  created_at:
                    anyOf:
                    - type: string
                    - type: 'null'
                    title: Created At
                  enabled:
                    title: Enabled
                    type: boolean
                  expires_at:
                    anyOf:
                    - type: string
                    - type: 'null'
                    title: Expires At
                  id:
                    title: Id
                    type: integer
                  key_prefix:
                    title: Key Prefix
                    type: string
                  last_used:
                    anyOf:
                    - type: string
                    - type: 'null'
                    title: Last Used
                  name:
                    title: Name
                    type: string
                  organization_id:
                    title: Organization Id
                    type: integer
                  rpm_limit:
                    anyOf:
                    - type: integer
                    - type: 'null'
                    title: Rpm Limit
                  tpm_limit:
                    anyOf:
                    - type: integer
                    - type: 'null'
                    title: Tpm Limit
                  usage:
                    $ref: '#/components/schemas/KeyDetailUsage'
                  user_id:
                    title: User Id
                    type: integer
                required:
                - id
                - name
                - key_prefix
                - user_id
                - organization_id
                - allowed_models
                - allowed_providers
                - budget_limit_daily
                - budget_limit_monthly
                - tpm_limit
                - rpm_limit
                - enabled
                - expires_at
                - last_used
                - created_at
                - usage
                title: KeyDetailResponse
                type: object
          description: Response body for GET /api/v1/keys/<key_id>.
      security:
      - bearerAuth: []
      summary: Get virtual key details
      tags:
      - Keys
    put:
      description: ''
      operationId: put_update_key
      parameters:
      - in: path
        name: key_id
        required: true
        schema:
          type: number
      requestBody:
        content:
          application/json:
            schema:
              description: Request body for PUT /api/v1/keys/<key_id>. Every field
                is a partial update.
              properties:
                allowed_models:
                  anyOf:
                  - items:
                      type: string
                    type: array
                  - type: 'null'
                  default: null
                  title: Allowed Models
                allowed_providers:
                  anyOf:
                  - items:
                      type: string
                    type: array
                  - type: 'null'
                  default: null
                  title: Allowed Providers
                budget_limit_daily:
                  anyOf:
                  - type: number
                  - type: 'null'
                  default: null
                  title: Budget Limit Daily
                budget_limit_monthly:
                  anyOf:
                  - type: number
                  - type: 'null'
                  default: null
                  title: Budget Limit Monthly
                enabled:
                  anyOf:
                  - type: boolean
                  - type: 'null'
                  default: null
                  title: Enabled
                expires_at:
                  anyOf:
                  - type: string
                  - type: 'null'
                  default: null
                  title: Expires At
                name:
                  anyOf:
                  - type: string
                  - type: 'null'
                  default: null
                  title: Name
                rpm_limit:
                  anyOf:
                  - type: integer
                  - type: 'null'
                  default: null
                  title: Rpm Limit
                tpm_limit:
                  anyOf:
                  - type: integer
                  - type: 'null'
                  default: null
                  title: Tpm Limit
              title: UpdateKeyRequest
              type: object
      responses:
        '200':
          content:
            application/json:
              schema:
                description: 'Generic `{"message": str}` envelope used by several
                  key endpoints.'
                properties:
                  message:
                    title: Message
                    type: string
                required:
                - message
                title: MessageResponse
                type: object
          description: 'Generic `{"message": str}` envelope used by several key endpoints.'
      security:
      - bearerAuth: []
      summary: Update virtual key
      tags:
      - Keys
  /api/v1/keys/{key_id}/rotate:
    post:
      description: ''
      operationId: post_rotate_key
      parameters:
      - in: path
        name: key_id
        required: true
        schema:
          type: number
      responses:
        '200':
          content:
            application/json:
              schema:
                description: Response body for POST /api/v1/keys/<key_id>/rotate.
                properties:
                  api_key:
                    title: Api Key
                    type: string
                  id:
                    title: Id
                    type: integer
                  key_prefix:
                    title: Key Prefix
                    type: string
                  message:
                    title: Message
                    type: string
                required:
                - id
                - api_key
                - key_prefix
                - message
                title: RotateKeyResponse
                type: object
          description: Response body for POST /api/v1/keys/<key_id>/rotate.
      security:
      - bearerAuth: []
      summary: Rotate key secret
      tags:
      - Keys
  /api/v1/keys/{key_id}/usage:
    get:
      description: ''
      operationId: get_get_key_usage
      parameters:
      - in: path
        name: key_id
        required: true
        schema:
          type: number
      responses:
        '200':
          content:
            application/json:
              schema:
                description: Response body for GET /api/v1/keys/<key_id>/usage.
                properties:
                  daily_usage:
                    items:
                      $ref: '#/components/schemas/KeyUsageDailyEntry'
                    title: Daily Usage
                    type: array
                  key_id:
                    title: Key Id
                    type: integer
                  key_name:
                    title: Key Name
                    type: string
                  period_days:
                    title: Period Days
                    type: integer
                  totals:
                    $ref: '#/components/schemas/KeyUsageTotals'
                required:
                - key_id
                - key_name
                - period_days
                - totals
                - daily_usage
                title: KeyUsageResponse
                type: object
          description: Response body for GET /api/v1/keys/<key_id>/usage.
      security:
      - bearerAuth: []
      summary: Get usage statistics for a key
      tags:
      - Keys
...